Customize Swaggerfile for Multiselections
Added Multiselection(boolean) for type=array parameters. In the interactive Docs
multiple values of the available values can be selected.
Enum key had to be moved one level up for this.

Added apiDefault(value) wich describes the used internal value if this
parameter is not specified.
As we are using the default key to prepopulate the inputs for the
required parameters, this is the way to show if it is a value that
is used by the API by default or an example value in the optional fields.